Extract ferrous-core crate and unify Platter logic (#2325)
- Created `crates/ferrous-core` to host shared `Platter` logic (2D scalar field).
- Refactored `experiments/ferrous-chimera`, `ferrous-graph`, `ferrous-neuron`, and `ferrous-fluid` to use `ferrous_core::Platter`.
- Standardized vector math in these experiments to use `locus::Vec2` and `f64`.
- Removed duplicate `platter.rs` files and local `Vec2` implementations.
- Added `accumulate` method to `Platter` to support unbounded density accumulation in `ferrous-fluid`.
- Verified all affected experiments compile and pass tests.

+++ experiments/ferrous-chimera/src/agent.rs
@@ -8,11 +8,11 @@ use ratatui::style::Color;
 pub struct Agent {
     pub body: Body,
     pub vm: ChimeraVM,
-    pub energy: f32,
+    pub energy: f64,
 }
 
 impl Agent {
-    pub fn new(dna: Dna, x: f32, y: f32) -> Self {
+    pub fn new(dna: Dna, x: f64, y: f64) -> Self {
         // Deterministic color from DNA length or content
         let hue = (dna.helix.strands.len() * 10) as u8;
         let color = Color::Rgb(hue, 255 - hue, 128);
@@ -24,31 +24,31 @@ impl Agent {
         }
     }
 
-    pub fn sense(&mut self, magnetism: f32) {
+    pub fn sense(&mut self, magnetism: f64) {
         // Push sensed magnetism to stack (normalized 0-1 mapped to 0-100 int)
         let sensor_val = (magnetism * 100.0) as i64;
         self.vm.stack.push(Value::Int(sensor_val));
     }
 
-    pub fn act(&mut self) -> (Vec2, f32) {
+    pub fn act(&mut self) -> (Vec2, f64) {
         // Interpret top of stack as action
         // Pop 2 values: X force, Y force.
         // Pop 1 value: Emit amount.
 
         let emit = if let Some(Value::Int(v)) = self.vm.stack.pop() {
-            (v as f32 / 100.0).clamp(0.0, 1.0)
+            (v as f64 / 100.0).clamp(0.0, 1.0)
         } else {
             0.0
         };
 
         let fy = if let Some(Value::Int(v)) = self.vm.stack.pop() {
-            (v as f32).clamp(-100.0, 100.0)
+            (v as f64).clamp(-100.0, 100.0)
         } else {
             0.0
         };
 
         let fx = if let Some(Value::Int(v)) = self.vm.stack.pop() {
-            (v as f32).clamp(-100.0, 100.0)
+            (v as f64).clamp(-100.0, 100.0)
         } else {
             0.0
         };
